+++ core/src/com/color/game/elements/staticelements/StaticPhysicComponent.java
@@ -78,7 +78,7 @@ public class StaticPhysicComponent extends PhysicComponent{
     }
 
     @Override
-    public void configureCircleBody(Vector2 position, int radius, World world, short category, short mask){
+    public void configureCircleBody(Vector2 position, float radius, World world, short category, short mask){
         this.world = world;
 
         this.bodyDef = new BodyDef();
